+++ include/NeoFOAM/core/executor/CPUExecutor.hpp
@@ -10,7 +10,8 @@ namespace NeoFOAM
 
 /**
  * @class CPUExecutor
- * @brief Reference executor for serial CPU execution.
+ * @brief Executor for handling multicore CPU based parallelization.
+ *
  *
  * @ingroup Executor
  */
@@ -18,7 +19,7 @@ class CPUExecutor
 {
 public:
 
-    using exec = Kokkos::Serial;
+    using exec = Kokkos::DefaultHostExecutionSpace;
 
     CPUExecutor();
     ~CPUExecutor();
@@ -35,6 +36,13 @@ public:
         return static_cast<T*>(Kokkos::kokkos_realloc<exec>(ptr, newSize * sizeof(T)));
     }
 
+    void* alloc(size_t size) const { return Kokkos::kokkos_malloc<exec>("Field", size); }
+
+    void* realloc(void* ptr, size_t newSize) const
+    {
+        return Kokkos::kokkos_realloc<exec>(ptr, newSize);
+    }
+
     /** @brief create a Kokkos view for a given ptr
      *
      * Based on the executor this function creates a Kokkos view into the data managed by ptr
@@ -46,14 +54,6 @@ public:
     decltype(auto) createKokkosView(ValueType* ptr, size_t size) const
     {
         return Kokkos::View<ValueType*, Kokkos::HostSpace, Kokkos::MemoryUnmanaged>(ptr, size);
-    }
-
-
-    void* alloc(size_t size) const { return Kokkos::kokkos_malloc<exec>("Field", size); }
-
-    void* realloc(void* ptr, size_t newSize) const
-    {
-        return Kokkos::kokkos_realloc<exec>(ptr, newSize);
     }
 
     std::string print() const { return std::string(exec::name()); }