+++ include/NeoFOAM/core/executor/executor.hpp
@@ -2,15 +2,15 @@
 // SPDX-FileCopyrightText: 2023 NeoFOAM authors
 #pragma once
 
-#include "NeoFOAM/core/executor/CPUExecutor.hpp"
+#include "NeoFOAM/core/executor/serialExecutor.hpp"
 #include "NeoFOAM/core/executor/GPUExecutor.hpp"
-#include "NeoFOAM/core/executor/OMPExecutor.hpp"
+#include "NeoFOAM/core/executor/CPUExecutor.hpp"
 #include <variant>
 
 namespace NeoFOAM
 {
 
-using Executor = std::variant<OMPExecutor, GPUExecutor, CPUExecutor>;
+using Executor = std::variant<CPUExecutor, GPUExecutor, SerialExecutor>;
 
 /**
  * @brief Checks if two executors are equal, i.e. they are of the same type.