+++ test/finiteVolume/cellCentred/boundary/volume/volFixedGradient.cpp
@@ -14,8 +14,8 @@
 TEST_CASE("fixedGradient")
 {
     NeoFOAM::Executor exec = GENERATE(
+        NeoFOAM::Executor(NeoFOAM::SerialExecutor {}),
         NeoFOAM::Executor(NeoFOAM::CPUExecutor {}),
-        NeoFOAM::Executor(NeoFOAM::OMPExecutor {}),
         NeoFOAM::Executor(NeoFOAM::GPUExecutor {})
     );